+++ examples/runnable/secret-basic/main.tf
@@ -3,7 +3,7 @@ terraform {
   required_providers {
     datahub = {
       source  = "datahub-project/datahub"
-      version = "0.6.0"
+      version = "0.7.0"
     }
   }
 }